Fix Obsoletes processing.
When A Obsoletes B, B is "auto removed to satisfy dependencies"
and A is "auto installed to satisfy dependencies".  If no package
depends on A, then A is "removed because no longer needed".

When new package installed, check if it Obsoletes other packages
and set "manual" flag if any of those packages were in "manual"
state before deletion.

Hopefully, this fully fixes dist-upgrade functionality.