ALT's pkgCacheFile::Open() has no WithLock parameter anymore; set it in ctor
Since a pkgCacheFile is a "oneshot, immutable" object, we can't have a
need to invoke Open() several times with different parameters.

(Unfortunately, this is a change in the API of ALT's APT. Here is a
justification for this change.

The old behavior of pkgCacheFile's methods in ALT's APT didn't match
the expectation of modern clients, such as PackageKit.

And when correcting the semantics of the pkgCacheFile class to match
the modern one, I rewrote it in a way that really guarantees that no
one will be able to change the owned pointers, and also I dared to
make it more simple and clear: the pointers to owned objects returned
from an instance of pkgCacheFile will simply never change during its
life; before, with Close(), the returned pointers would be new after
Close(), which is not really needed in actual applications, as we see.

After the rewrite, the use of pkgCacheFile becomes alike to the common
RAII idiom[1], with the specific feature that the resources are acquired
lazily after construction.

Open() has been left as a way to force the acquisition of all
"resources", i.e., the build of all caches, and to check errors during it.

[1]: https://en.cppreference.com/w/cpp/language/raii "Classes with
open()/close(), lock()/unlock(), or init()/copyFrom()/destroy() member
functions are typical examples of non-RAII classes"

)