threads.h: put missing #include <pthread.h>, necessary for private members
Before, it was implicitly included through stdc++ headers (up to gcc-12.1.1-alt2),
but that has been fixed in gcc-12.2.1-alt1.

Unfortunately, although it is needed only for the private
implementation of the classes in threads.h, there is no way to make
"#include <pthread.h>" private, so as not to pollute those who use
threads.h (because of the private data members, which are defined in
here).

Hiding the include in namespace threads is also a bad idea, because
this would break everything around badly (any code which needs to
include <pthread.h>, too, after threads.h).